+++ src/program-instructions.ts
@@ -1755,6 +1755,25 @@ export function overrideExpiryIx(
   });
 }
 
+export async function toggleMarketMakerIx(
+  isMarketMaker: boolean,
+  zetaGroup: PublicKey,
+  user: PublicKey
+): Promise<TransactionInstruction> {
+  let [marginAccount, _nonce] = await utils.getMarginAccount(
+    Exchange.programId,
+    zetaGroup,
+    user
+  );
+  return Exchange.program.instruction.toggleMarketMaker(isMarketMaker, {
+    accounts: {
+      state: Exchange.stateAddress,
+      admin: Exchange.state.admin,
+      marginAccount,
+    },
+  });
+}
+
 export interface ExpireSeriesOverrideArgs {
   settlementNonce: number;
   settlementPrice: anchor.BN;
@@ -1788,6 +1807,7 @@ export interface StateParams {
   nativeDepositLimit: anchor.BN;
   expirationThresholdSeconds: number;
   positionMovementFeeBps: number;
+  marginConcessionPercentage: number;
 }
 
 export interface UpdatePricingParametersArgs {